Modified installers for desktop applications to support multiple users.

+++ provisioners/scripts/centos/install_centos7_desktop_users_env.sh
@@ -4,6 +4,9 @@
 # set default value for devops home environment variable if not set. -----------
 devops_home="${devops_home:-/opt/devops}"                   # [optional] devops home (defaults to '/opt/devops').
 
+# set current date for temporary filename. -------------------------------------
+curdate=$(date +"%Y-%m-%d.%H-%M-%S")
+
 # create default environment profile for user 'root'. --------------------------
 rootprofile="${devops_home}/provisioners/scripts/common/users/user-root-bash_profile.sh"
 rootrc="${devops_home}/provisioners/scripts/common/users/user-root-bashrc.sh"
@@ -20,8 +23,18 @@ sed -i 's/^#alias gvim/alias gvim/g' ${rootrc}
 
 # copy environment profiles to user 'root' home.
 cd /root
-cp -p .bash_profile .bash_profile.orig
-cp -p .bashrc .bashrc.orig
+
+if [ -e ".bash_profile.orig" ]; then
+  cp -p .bash_profile .bash_profile.${curdate}
+else
+  cp -p .bash_profile .bash_profile.orig
+fi
+
+if [ -e ".bashrc.orig" ]; then
+  cp -p .bashrc .bashrc.${curdate}
+else
+  cp -p .bashrc .bashrc.orig
+fi
 
 cp -f ${rootprofile} .bash_profile
 cp -f ${rootrc} .bashrc
@@ -58,8 +71,18 @@ sed -i 's/^#alias gvim/alias gvim/g' ${vagrantrc}
 
 # copy environment profiles to user 'vagrant' home.
 cd /home/vagrant
-cp -p .bash_profile .bash_profile.orig
-cp -p .bashrc .bashrc.orig
+
+if [ -e ".bash_profile.orig" ]; then
+  cp -p .bash_profile .bash_profile.${curdate}
+else
+  cp -p .bash_profile .bash_profile.orig
+fi
+
+if [ -e ".bashrc.orig" ]; then
+  cp -p .bashrc .bashrc.${curdate}
+else
+  cp -p .bashrc .bashrc.orig
+fi
 
 cp -f ${vagrantprofile} .bash_profile
 cp -f ${vagrantrc} .bashrc
@@ -102,4 +125,5 @@ chmod 644 ${userfolder}/${dccompletion_binary}
 # configure gnome-3 desktop properties for devops users. -----------------------
 cd ${devops_home}/provisioners/scripts/centos
 chmod 755 config_centos7_gnome_desktop.sh
-runuser -c "${devops_home}/provisioners/scripts/centos/config_centos7_gnome_desktop.sh" - vagrant
+#sudo -u vagrant -S -E sh -eux "${devops_home}/provisioners/scripts/centos/config_centos7_gnome_desktop.sh"
+runuser -c "devops_home=${devops_home} ${devops_home}/provisioners/scripts/centos/config_centos7_gnome_desktop.sh" - vagrant
